requirements.txt
fastap
uvicorn[standard]
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose
pydantic
langchain
langgraph
pydantic_ai
psycopg2-binary
PyJWT
urllib3
